requirements/base.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-cloud-storage==3.2.0
PyPDF2==3.0.1
PyMuPDF==1.26.3
python-dotenv==1.1.1
anthropic==0.60.0
python-multipart==0.0.20
slowapi==0.1.9
